refactor: apply code covention to code related post
refer to naver code convention
