+++ src/main/java/com/moyeota/moyeotaproject/controller/ParticipationDetailsController.java
@@ -6,7 +6,7 @@ import com.moyeota.moyeotaproject.config.exception.ApiException;
 import com.moyeota.moyeotaproject.config.exception.ErrorCode;
 import com.moyeota.moyeotaproject.controller.dto.participationDetailsDto.DistancePriceDto;
 import com.moyeota.moyeotaproject.controller.dto.participationDetailsDto.ParticipationDetailsResponseDto;
-import com.moyeota.moyeotaproject.controller.dto.postsDto.PostsResponseDto;
+import com.moyeota.moyeotaproject.controller.dto.postsdto.PostsGetResponseDto;
 import com.moyeota.moyeotaproject.domain.participationDetails.ParticipationDetails;
 import com.moyeota.moyeotaproject.domain.participationDetails.ParticipationDetailsStatus;
 import com.moyeota.moyeotaproject.domain.posts.PostsStatus;
@@ -40,7 +40,7 @@ public class ParticipationDetailsController {
     @PostMapping("/join/posts/{postId}")
     public ResponseDto<Long> join(HttpServletRequest request,
                                   @ApiParam(value = "모집글 인덱스 번호") @PathVariable("postId") Long postId) {
-        PostsResponseDto responseDto = postsService.findById(postId);
+        PostsGetResponseDto responseDto = postsService.findById(postId);
         if (responseDto.getNumberOfParticipants() == responseDto.getNumberOfRecruitment()) {
             throw new ApiException(ErrorCode.POSTS_ALREADY_FINISH);
         } else if (responseDto.getStatus() == PostsStatus.COMPLETE) {
@@ -77,7 +77,7 @@ public class ParticipationDetailsController {
     //내가 참가 신청한 모집글 목록 조회 API (최신순으로)
     @ApiOperation(value = "특정 회원의 참가 신청 모집글 전체 조회", notes = "특정 회원이 참가 신청한 모집글을 전체 조회하는 API")
     @GetMapping("/")
-    public ResponseDto<List<PostsResponseDto>> findMyParticipationDetailsDesc(HttpServletRequest request) {
+    public ResponseDto<List<PostsGetResponseDto>> findMyParticipationDetailsDesc(HttpServletRequest request) {
         return ResponseUtil.SUCCESS("모집글 조회에 성공하였습니다.",
                 participationDetailsService.findMyParticipationDetailsDesc(request.getHeader("Authorization")));
     }
