--- src/main/java/com/moyeota/moyeotaproject/domain/chatMessage/MessageRoomIdMapping.java
@@ -1,5 +1,0 @@
-package com.moyeota.moyeotaproject.domain.chatMessage;
-
-public interface MessageRoomIdMapping {
-    String getRoomId();
-}
