--- src/main/java/com/moyeota/moyeotaproject/domain/chatMessage/MessageStatus.java
@@ -1,5 +1,0 @@
-package com.moyeota.moyeotaproject.domain.chatMessage;
-
-public enum MessageStatus {
-    VALID, NON_VALID
-}
