--- src/main/java/com/moyeota/moyeotaproject/domain/chatMessage/MessageType.java
@@ -1,5 +1,0 @@
-package com.moyeota.moyeotaproject.domain.chatMessage;
-
-public enum MessageType {
-    ENTER, TALK, LEAVE
-}
