+++ src/main/java/com/moyeota/moyeotaproject/domain/posts/Category.java
@@ -1,5 +1,5 @@
 package com.moyeota.moyeotaproject.domain.posts;
 
 public enum Category {
-    COLLEGE, LIFE, COMMUTE, RESERVIST;
+	COLLEGE, LIFE, COMMUTE, RESERVIST;
 }
