+++ src/main/java/com/moyeota/moyeotaproject/domain/posts/PostsStatus.java
@@ -1,5 +1,5 @@
 package com.moyeota.moyeotaproject.domain.posts;
 
 public enum PostsStatus {
-    RECRUITING, COMPLETE
+	RECRUITING, COMPLETE
 }
