+++ src/main/java/com/moyeota/moyeotaproject/domain/posts/SameGender.java
@@ -1,5 +1,5 @@
 package com.moyeota.moyeotaproject.domain.posts;
 
 public enum SameGender {
-    YES, NO
+	YES, NO
 }
