+++ src/main/java/com/moyeota/moyeotaproject/domain/posts/Vehicle.java
@@ -1,5 +1,5 @@
 package com.moyeota.moyeotaproject.domain.posts;
 
 public enum Vehicle {
-    일반, 밴
+	일반, 밴
 }
